+++ tests/components/device_tracker/test_init.py
@@ -13,7 +13,7 @@ from homeassistant.core import callback, State
 from homeassistant.setup import setup_component, async_setup_component
 from homeassistant.helpers import discovery
 from homeassistant.loader import get_component
-from homeassistant.util.async import run_coroutine_threadsafe
+from homeassistant.util.async_ import run_coroutine_threadsafe
 import homeassistant.util.dt as dt_util
 from homeassistant.const import (
     ATTR_ENTITY_ID, ATTR_ENTITY_PICTURE, ATTR_FRIENDLY_NAME, ATTR_HIDDEN,
