+++ tests/components/media_player/test_async_helpers.py
@@ -5,7 +5,7 @@ import asyncio
 import homeassistant.components.media_player as mp
 from homeassistant.const import (
     STATE_PLAYING, STATE_PAUSED, STATE_ON, STATE_OFF, STATE_IDLE)
-from homeassistant.util.async import run_coroutine_threadsafe
+from homeassistant.util.async_ import run_coroutine_threadsafe
 
 from tests.common import get_test_home_assistant
