+++ tests/components/switch/test_rest.py
@@ -5,7 +5,7 @@ import aiohttp
 
 import homeassistant.components.switch.rest as rest
 from homeassistant.setup import setup_component
-from homeassistant.util.async import run_coroutine_threadsafe
+from homeassistant.util.async_ import run_coroutine_threadsafe
 from homeassistant.helpers.template import Template
 from tests.common import get_test_home_assistant, assert_setup_component
