+++ tests/helpers/test_aiohttp_client.py
@@ -8,7 +8,7 @@ import pytest
 from homeassistant.core import EVENT_HOMEASSISTANT_CLOSE
 from homeassistant.setup import async_setup_component
 import homeassistant.helpers.aiohttp_client as client
-from homeassistant.util.async import run_callback_threadsafe
+from homeassistant.util.async_ import run_callback_threadsafe
 
 from tests.common import get_test_home_assistant
